requirements.txt
appdirs==1.4.4
asgiref==3.2.10
attrs==19.3.0
black==19.10b0
cfgv==3.2.0
click==7.1.2
distlib==0.3.1
Django==3.1
django-filter==2.3.0
djangorestframework==3.11.0
filelock==3.0.12
flake8==3.8.3
identify==1.4.25
importlib-metadata==1.7.0
importlib-resources==3.0.0
Markdown==3.2.2
mccabe==0.6.1
nodeenv==1.4.0
pathspec==0.8.0
pre-commit==2.6.0
pycodestyle==2.6.0
pyflakes==2.2.0
pytz==2020.1
PyYAML==5.3.1
regex==2020.7.14
six==1.15.0
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
virtualenv==20.0.30
zipp==3.1.0
